+++ apps/commerce-client-web/src/app/(default)/order/page.tsx
@@ -1,105 +1,34 @@
-'use client';
-
-import React, { useState, useEffect } from 'react';
-import PaymentProducts from './components/payment-products';
-import PaymentSummary from './components/payment-summary';
-import OrderShippingInfo from './components/payment-shippinginfo';
-import OrderPaymentMethod from './components/payment-method';
-import PaymentAgreement from './components/payment-agreement';
-import PaymentUserInfo from './components/payment-userInfo';
-import useCartStore from '@/stores/use-cart-store';
+import PaymentPageClient from '@/app/(default)/order/components/payment-page-client';
+import mswApi from '@/lib/msw-api';
 import { CartItem } from '@/types/cart-types';
 import { DetailBook } from '@/types/book-types';
-import mswApi from '@/lib/msw-api';
 
-const mockUser = {
-  name: '유재석',
-  phnum: '01024129368',
-  email: 'ddd@naver.com',
-};
+interface PaymentPageProps {
+  searchParams: { productId?: string };
+}
 
-const mockOrder = {
-  name: '패스트파이브',
-  phnum: '01024129368',
-  address: '서울',
-  detailAddress: '강남구',
-  memo: '놓고 가주세여',
-};
+const PaymentPage = async ({ searchParams }: PaymentPageProps) => {
+  let books: CartItem[] = [];
 
-const PaymentPage = () => {
-  const { getSelectedBook, items } = useCartStore();
-  const [books, setBooks] = useState<CartItem[]>([]);
-  const [user, setUser] = useState(mockUser);
-  const [order, setOrder] = useState(mockOrder);
-
-  useEffect(() => {
-    const fetchBookData = async () => {
-      const searchParams = new URLSearchParams(window.location.search);
-      const productId = searchParams.get('productId');
-
-      if (productId) {
-        const bookId = Number(productId);
-        try {
-          const book = await mswApi(`books/${bookId}`).json<DetailBook>();
-
-          if (book) {
-            const cartItem: CartItem = {
-              ...book,
-              selectNum: 1,
-              selected: true,
-              shippingInfo: '무료배송',
-              imageUrl: book.coverImage || undefined,
-            };
-
-            setBooks([cartItem]);
-          }
-        } catch (error) {
-          console.error('Failed to fetch book:', error);
-        }
-      } else {
-        setBooks(getSelectedBook());
+  if (searchParams.productId) {
+    const bookId = Number(searchParams.productId);
+    try {
+      const book = await mswApi(`books/${bookId}`).json<DetailBook>();
+      if (book) {
+        const cartItem: CartItem = {
+          ...book,
+          selectNum: 1,
+          selected: true,
+          shippingInfo: '무료배송',
+          imageUrl: book.coverImage || undefined,
+        };
+        books = [cartItem];
       }
-    };
-
-    fetchBookData();
-  }, [getSelectedBook, items]);
-
-  const handleUserChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setUser((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleOrderChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target;
-    setOrder((prev) => ({ ...prev, [name]: value }));
-  };
-
-  return (
-    <div className="container mx-auto p-6">
-      <h1 className="mb-8 text-center text-2xl font-semibold">결제하기</h1>
-
-      {/* 그리드 설정: 기본 1열, md(768px) 이상 2열 */}
-      <div className="grid gap-6 md:grid-cols-1 lg:grid-cols-2">
-        {/* 주문 상품 정보 */}
-        <PaymentProducts books={books} />
-
-        {/* 주문 요약 */}
-        <PaymentSummary books={books} />
-
-        {/* 주문자 정보 */}
-        <PaymentUserInfo user={user} onUserChange={handleUserChange} />
-
-        {/* 배송 정보 */}
-        <OrderShippingInfo order={order} onOrderChange={handleOrderChange} />
-
-        {/* 결제수단 */}
-        <OrderPaymentMethod />
-
-        {/* 동의 및 결제하기 */}
-        <PaymentAgreement />
-      </div>
-    </div>
-  );
+    } catch (error) {
+      console.error('Failed to fetch book:', error);
+    }
+  }
+  return <PaymentPageClient serverBooks={books} />;
 };
 
 export default PaymentPage;
